requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
parameterized==0.8.1
pybson==0.5.9
pymongo==3.12.1
python-dateutil==2.8.2
requests==2.26.0
schedule==1.1.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
